fix(fibonacci/fibonacci.cpp): correct typing

+++ src/fibonacci/fibonacci.cpp
@@ -1,6 +1,6 @@
 #include <nan.h>
 
-double fib(double n) {
+int fib(int n) {
     if (n == 0 || n == 1)
         return n;
     else
